Fix formatting regressions
Something like

CARGO_USE_GITHUB=

would lead to a crash due to an unsigned wrap around because
16 < strlen("CARGO_USE_GITHUB=")

Also fix an off by one for things like LICENSE_FILE_GPLv3+ which
need an extra space before the modifier.  This was fine before the
AST changes.

+++ ast.h
@@ -196,7 +196,7 @@ struct ASTNode {
 	struct ASTNodeLineRange line_end;
 	int edited;
 	struct {
-		int goalcol;
+		size_t goalcol;
 	} meta;
 	union {
 		struct ASTNodeRoot root;
